backend/requirements.txt
python-dotenv
pydantic
fastapi
uvicorn
langchain
langchain-google-genai
langchain-core
langchain-community
langchain-text-splitters
chromadb